Shift FileInfoGenerator intoFileInfoExtractor and remove PMD
