+++ src/main/java/reposense/authorship/FileInfoExtractor.java
@@ -1,5 +1,6 @@
 package reposense.authorship;
 
+import java.io.BufferedReader;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -12,8 +13,8 @@ import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import reposense.authorship.extractor.FileInfoGenerator;
 import reposense.authorship.model.FileInfo;
+import reposense.authorship.model.LineInfo;
 import reposense.git.GitChecker;
 import reposense.model.RepoConfiguration;
 import reposense.system.LogsManager;
@@ -58,13 +59,31 @@ public class FileInfoExtractor {
                 }
 
                 if (relativePath.endsWith(".java") || relativePath.endsWith(".adoc")) {
-                    fileInfos.add(FileInfoGenerator.generateFileInfo(
-                            config.getRepoRoot(), relativePath.replace('\\', '/')));
+                    fileInfos.add(generateFileInfo(config.getRepoRoot(), relativePath.replace('\\', '/')));
                 }
             }
         } catch (IOException ioe) {
             logger.log(Level.SEVERE, "Error occured while extracing all relevant file infos.", ioe);
         }
+    }
+
+    /**
+     * Generates and returns a {@code FileInfo} with a list of {@code LineInfo} for each line content in the
+     * {@code relativePath} file.
+     */
+    public static FileInfo generateFileInfo(String repoRoot, String relativePath) {
+        FileInfo fileInfo = new FileInfo(relativePath);
+        Path path = Paths.get(repoRoot, relativePath);
+        try (BufferedReader br = new BufferedReader(Files.newBufferedReader(path))) {
+            String line;
+            int lineNum = 1;
+            while ((line = br.readLine()) != null) {
+                fileInfo.addLine(new LineInfo(lineNum++, line));
+            }
+        } catch (IOException ioe) {
+            logger.log(Level.SEVERE, ioe.getMessage(), ioe);
+        }
+        return fileInfo;
     }
 
     /**
